requirements.txt
fastapi
uvicorn
transformers
torch
pyjwt
